requirements.txt
# Multi-Agent Code Generation System Dependencies
openai>=1.0.0
anthropic>=0.7.0
pydantic>=2.0.0
asyncio-throttle>=1.0.2
aiohttp>=3.9.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
jinja2>=3.1.2
pyyaml>=6.0.1
httpx>=0.25.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
